requirements.txt
flask
ultralytics
torch
pillow
